Extract ResourceApplicator and CrewReservationManager from ParsekScenario
- ResourceApplicator: TickStandalone, TickTrees, DeductBudget, CorrectToBaseline
  moved from ParsekScenario (coroutine shells stay, mutation logic moves)
- CrewReservationManager: crewReplacements dict + all reservation methods
  (Reserve/Unreserve/Swap/Clear/Extract/ShouldProcess/ShouldRemoveEva)
  moved from ParsekScenario, ~40 call sites updated across 7 source files
- CrewLog helper preserves ScenarioLog WARNING routing behavior
- ParsekFlight resource code untouched (different logging/guards)
- ParsekScenario reduced by ~720 lines, all 3325 tests pass
