+++ Source/Parsek.Tests/GroupManagementTests.cs
@@ -20,7 +20,7 @@ namespace Parsek.Tests
             ParsekLog.SuppressLogging = false;
             ParsekLog.TestSinkForTesting = line => logLines.Add(line);
             GroupHierarchyStore.ResetGroupsForTesting();
-            ParsekScenario.ResetReplacementsForTesting();
+            CrewReservationManager.ResetReplacementsForTesting();
         }
 
         public void Dispose()
@@ -30,7 +30,7 @@ namespace Parsek.Tests
             RecordingStore.SuppressLogging = true;
             RecordingStore.ResetForTesting();
             GroupHierarchyStore.ResetGroupsForTesting();
-            ParsekScenario.ResetReplacementsForTesting();
+            CrewReservationManager.ResetReplacementsForTesting();
         }
 
         // ─── IsInAncestorChain ──────────────────────────────────────────────
